# apps/txd/ubc-app/src/_assets/api.yml
swagger: '2.0'
info: 
  description: UBC backend API
  version: '1.0.0'
  title: 'UBC API'
basePath: '/api/v1'
schemes:
  - http
  - https
...
paths:
  /image/{id}:
    get:
      operationId: 'getImage'
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
    post:
      operationId: 'uploadImage'
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
...